requirements.txt
numpy==1.15.1
websocket-client==0.51.0
boto3==1.8.3
botocore==1.11.3
ujson==1.35
awscli==1.16.3
raven==6.9.0
requests[security]==2.19.1
signalr-client-threads==0.0.11
urllib3==1.22
